add uninstall

+++ install.sh
@@ -8,12 +8,24 @@ if [ ! -w $PREFIX/lib ] || [ ! -w $PREFIX/include ];
 then
 	PREFIX=~/.ft
 fi
-echo "Install in prefix : $PREFIX"
-mkdir -p $PREFIX/lib
-mkdir -p $PREFIX/include
-echo "Install librarys ..."
-mv $1 $PREFIX/lib
-mkdir -p $PREFIX/$2
-echo "Install includes ..."
-cp $2/common.h $PREFIX/$2
-rm -rf $3
+
+if [ ! -z $4 ] && [ "$4" = "uninstall" ];
+then
+	echo "Uninstall in prefix : $PREFIX"
+	echo "Removing librarys ..."
+	rm -f $PREFIX/$1
+	echo "Removing includes ..."
+	rm -f $PREFIX/$2/common.h
+else
+	if [ ! -z $1 ] && [ ! -z $2 ] && [ ! -z $3 ];
+	then
+		echo "Install in prefix : $PREFIX"
+		echo "Install librarys ..."
+		mkdir -p $PREFIX/`dirname $1`
+		mv $1 $PREFIX/`dirname $1`
+		mkdir -p $PREFIX/$2
+		echo "Install includes ..."
+		cp $2/common.h $PREFIX/$2
+		rm -rf $3
+	fi
+fi
